refactor: extract hotkey parsing into services/hotkey.ts
Pure functions parseHotkey(), matchesHotkey(), KEY_NAME_TO_CODES map,
and ParsedHotkey type extracted from main/index.ts into
main/services/hotkey.ts. These are stateless and testable in isolation.

Further subsystem extraction (volume, tray, windows) deferred —
these are tightly coupled to shared mutable state and need a state
management layer before they can be cleanly extracted.
